requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
Django==5.1.7
django-environ==0.12.0
django-filter==25.1
django-graphql-jwt==0.4.0
djangorestframework==3.15.2
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
pillow==11.1.0
promise==2.3
psycopg2==2.9.10
PyJWT==2.10.1
Rx==1.6.3
singledispatch==4.1.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
